render resolution options "doors" and 3D world objects water edge & over-water object fixes window save/restore improve particle shader set equipment UI weapon model embedded fx proper implementation +lots of stuff

+++ Sacred.Core/World/Sector/SectorEnvironmentFlags.cs
@@ -5,7 +5,7 @@ namespace Sacred.Core.World.Sector;
 /// The boundary flags identify an absent neighbor on the corresponding map edge.
 /// </summary>
 [Flags]
-public enum SectorEnvironmentFlags : byte
+public enum SectorEnvironmentFlags : uint
 {
     None = 0,
     /// <summary>Unresolved flag bit. The name preserves its raw hexadecimal byte value.</summary>
